fix(apis/v2/inner): split esb inner apis into separate files, for multi-tenant
